requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pokerkit
websockets